In preparation of deployment and config update, and some optimisation
