+++ .env.example
@@ -1,15 +1,15 @@
-AWS_ACCESS_KEY_ID="43242s387fg44638s244fh24"
-AWS_REGION="us-east-2"
-AWS_S3_BUCKET="hackathon-backups"
-AWS_S3_ENDPOINT="https://dfh7346398578dhgsds.r2.cloudflarestorage.com"
-AWS_S3_REGION="auto"
-AWS_SECRET_ACCESS_KEY="4258722gyrgejw78gfg2u20974fhihdfs02"
-AWS_SES_ACCESS_KEY="ADEIAFHJ3HKFU48FHJK"
-AWS_SES_EMAIL_FROM="no-reply@[your_organization].org"
-AWS_SES_SECRET_ACCESS_KEY="AfBHKrDFh/Q03u+37iyGrwjFJ"
-BACKUP_CRON_SCHEDULE="0 */6 * * *"
-BACKUP_DATABASE_URL="postgres://default:RtN4q5qefsTO@ui-painted-hills-e5ddlb.us-east-2.aws.neon.tech:4576/verceldb?sslmode=require"
-BLOB_READ_WRITE_TOKEN="vercel_blob_rw_fhb9DH2kOEIO3EOI_3dBKKFHBGDLNSjnjDOIJHF"
+# AWS_ACCESS_KEY_ID="43242s387fg44638s244fh24"
+# AWS_REGION="us-east-2"
+# AWS_S3_BUCKET="hackathon-backups"
+# AWS_S3_ENDPOINT="https://dfh7346398578dhgsds.r2.cloudflarestorage.com"
+# AWS_S3_REGION="auto"
+# AWS_SECRET_ACCESS_KEY="4258722gyrgejw78gfg2u20974fhihdfs02"
+# AWS_SES_ACCESS_KEY="ADEIAFHJ3HKFU48FHJK"
+# AWS_SES_EMAIL_FROM="no-reply@[your_organization].org"
+# AWS_SES_SECRET_ACCESS_KEY="AfBHKrDFh/Q03u+37iyGrwjFJ"
+# BACKUP_CRON_SCHEDULE="0 */6 * * *"
+# BACKUP_DATABASE_URL="postgres://default:RtN4q5qefsTO@ui-painted-hills-e5ddlb.us-east-2.aws.neon.tech:4576/verceldb?sslmode=require"
+# BLOB_READ_WRITE_TOKEN="vercel_blob_rw_fhb9DH2kOEIO3EOI_3dBKKFHBGDLNSjnjDOIJHF"
 BOT_API_URL="https://hackkit-actions-bots.up.railway.app"
 CLERK_SECRET_KEY="sk_test_AFNJKABF2uhkbDBUKAFAKBAHBj3r"
 CLOUDFLARE_ACCOUNT_ID="cdcdJB3453KKJFBDSBHA54546UIBF"
@@ -19,21 +19,21 @@ DISCORD_DEV_VERIFY_CHANNEL_ID="52872856278050172988890"
 DISCORD_PROD_SERVER_ID="528728562780501729853789"
 DISCORD_PROD_VERIFY_CHANNEL_ID="5287378262780501729853789"
 DISCORD_SECRET_TOKEN="FHBAHJABDhbHSUeyi7398.S72ljd.HKBVAfug73gifa-74gfwiyB-BHJBDHV07hAJf83"
-HK_ENV="development"
+NODE_ENV="development"
 INTERNAL_AUTH_KEY="bwgybgidbsi-4784gyfgs-475hhkdsbfs-bwgybgidbsi-4784gyfgs-475hhkdsbfs-bwgybgidbsi-4784gyfgs-475hhkdsbfs-fbnauib4783-bhfbsjfbs"
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY="pk_test_YRGIBHSBIbabffjdhvbuYGI7BK"
 NEXT_PUBLIC_CLERK_SIGN_IN_URL="/sign-in"
 NEXT_PUBLIC_CLERK_SIGN_UP_URL="/sign-up"
-PLUNK_API_KEY="sk_279y482gibfai7g74gikbga345"
-PLUNK_API_URL="https://plunk.[your_org_name].org/api/[version_number]"
-POSTGRES_DATABASE="verceldb"
-POSTGRES_HOST="ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech"
-POSTGRES_PASSWORD="dkhjbIYBFABUI"
-POSTGRES_PRISMA_URL="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?pgbouncer=true&connect_timeout=15&sslmode=require"
-POSTGRES_URL="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?sslmode=require"
-POSTGRES_URL_NO_SSL="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb"
-POSTGRES_URL_NON_POOLING="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?sslmode=require"
-POSTGRES_USER="default"
+# PLUNK_API_KEY="sk_279y482gibfai7g74gikbga345"
+# PLUNK_API_URL="https://plunk.[your_org_name].org/api/[version_number]"
+# POSTGRES_DATABASE="verceldb"
+# POSTGRES_HOST="ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech"
+# POSTGRES_PASSWORD="dkhjbIYBFABUI"
+# POSTGRES_PRISMA_URL="postgres://default:dkhjbIYBFABUI@ep-rising-pfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?pgbouncer=true&connect_timeout=15&sslmode=require"
+# POSTGRES_URL="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?sslmode=require"
+# POSTGRES_URL_NO_SSL="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb"
+# POSTGRES_URL_NON_POOLING="postgres://default:dkhjbIYBFABUI@ep-rising-sky-adbkfahbeke-spooler.us-east-2.aws.neon.tech/verceldb?sslmode=require"
+# POSTGRES_USER="default"
 R2_ACCESS_KEY_ID="1bfkhbfyiayi38uhidfis"
 R2_BUCKET_NAME="your-org-name-userdata"
 R2_SECRET_ACCESS_KEY="279y45g79tgbjsbfhbsiufhbs89hg487hsiufs"
