+++ apps/web/.gitignore
@@ -36,3 +36,4 @@ next-env.d.ts
 
 # react-email
 .react-email/
+certificates
