+++ apps/web/package.json
@@ -11,10 +11,8 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.750.0",
-    "@aws-sdk/client-ses": "^3.616.0",
     "@aws-sdk/s3-request-presigner": "^3.750.0",
     "@clerk/nextjs": "^6.12.12",
-    "@gsap/react": "^2.1.1",
     "@hookform/resolvers": "^3.9.0",
     "@internationalized/date": "^3.5.4",
     "@planetscale/database": "^1.18.0",
@@ -32,22 +30,9 @@
     "@react-email/components": "^0.0.21",
     "@t3-oss/env-nextjs": "^0.10.1",
     "@tanstack/match-sorter-utils": "^8.19.4",
-    "@tanstack/react-query": "^5.51.11",
     "@tanstack/react-table": "^8.19.3",
-    "@trpc/client": "11.0.0-rc.466",
-    "@trpc/next": "11.0.0-rc.466",
-    "@trpc/react-query": "11.0.0-rc.466",
-    "@trpc/server": "11.0.0-rc.466",
-    "@types/node": "20.14.11",
-    "@types/react": "18.3.3",
-    "@types/react-dom": "18.3.0",
     "@upstash/redis": "^1.34.4",
-    "@vercel/analytics": "^1.3.1",
-    "@vercel/blob": "^0.10.0",
-    "@vercel/postgres": "^0.9.0",
     "@yudiel/react-qr-scanner": "^2.1.0",
-    "autoprefixer": "10.4.19",
-    "axios": "^1.7.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
@@ -55,32 +40,21 @@
     "date-fns": "^3.6.0",
     "date-fns-tz": "^3.1.3",
     "db": "workspace:*",
-    "dotenv": "^16.4.5",
     "drizzle-orm": "^0.39.3",
     "drizzle-zod": "^0.7.0",
     "embla-carousel": "8.1.7",
-    "embla-carousel-autoplay": "8.1.7",
     "embla-carousel-react": "8.1.7",
-    "framer-motion": "^11.3.8",
-    "gsap": "^3.12.5",
-    "jiti": "^1.21.6",
     "lucide-react": "^0.411.0",
     "nanoid": "^5.0.7",
     "next": "14.2.25",
     "next-safe-action": "^7.9.3",
     "nextjs@4": "link:clerk/nextjs@4",
     "no-profanity": "^1.5.1",
-    "pg": "^8.12.0",
-    "postcss": "8.4.39",
-    "postgres": "^3.4.4",
-    "pusher-js": "8.4.0-rc2",
     "react": "18.3.1",
     "react-aria": "^3.33.1",
     "react-confetti": "^6.1.0",
     "react-dom": "18.3.1",
     "react-dropzone": "^14.2.3",
-    "react-email": "^2.1.5",
-    "react-fast-marquee": "^1.6.5",
     "react-hook-form": "7.52.1",
     "react-loader-spinner": "^6.1.6",
     "react-parallax-tilt": "^1.7.232",
@@ -92,9 +66,6 @@
     "sonner": "^1.5.0",
     "superjson": "^2.2.1",
     "tailwind-merge": "^2.4.0",
-    "tailwindcss": "3.4.6",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.5.3",
     "use-debounce": "^10.0.1",
     "usehooks-ts": "^3.1.0",
     "vaul": "^0.9.1",
@@ -102,6 +73,16 @@
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.13.10",
-    "esbuild-register": "^3.5.0"
+    "@types/node": "20.14.11",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "autoprefixer": "10.4.19",
+    "dotenv": "16.4.5",
+    "esbuild-register": "^3.5.0",
+    "jiti": "1.21.6",
+    "postcss": "8.4.39",
+    "tailwindcss": "3.4.6",
+    "tailwindcss-animate": "1.0.7",
+    "typescript": "5.5.3"
   }
 }
