--- apps/web/src/app/admin/points/page.tsx
@@ -1,7 +1,0 @@
-export default function Points() {
-	return (
-		<main className="h-dvh pt-28">
-			<p>Coming soon...</p>
-		</main>
-	);
-}
