--- apps/web/src/app/admin/toggles/tickets/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-	return <p>Coming soon</p>;
-}
