--- apps/web/src/app/dash/tickets/page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-	return (
-		<main className="flex h-full items-center justify-center">
-			No Ticket Selected
-		</main>
-	);
-}
