+++ apps/web/src/app/layout.tsx
@@ -1,8 +1,7 @@
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
 import { cookies } from "next/headers";
-import { Analytics } from "@vercel/analytics/react";
-import { defaultTheme } from "config";
+import { defaultTheme, themeTokens } from "config";
 
 export default function RootLayout({
 	children,
@@ -10,12 +9,21 @@ export default function RootLayout({
 	children: React.ReactNode;
 }) {
 	const theme = cookies().get("hk_theme")?.value || defaultTheme;
+
+	const buildVars = (map: Record<string, string>) =>
+		Object.entries(map)
+			.map(([k, v]) => `--${k}: ${v};`)
+			.join("\n");
+
+	const cssVars = `:root {\n${buildVars(themeTokens.light)}\n}\n.dark {\n${buildVars(
+		themeTokens.dark as Record<string, string>,
+	)}\n}`;
 	return (
 		<ClerkProvider>
 			<html lang="en">
 				<body className={theme === "dark" ? "dark" : ""}>
+					<style dangerouslySetInnerHTML={{ __html: cssVars }} />
 					{children}
-					<Analytics />
 				</body>
 			</html>
 		</ClerkProvider>
