--- apps/web/src/emails/RegistrationSuccessEmail.tsx
@@ -1,114 +1,0 @@
-import {
-	Body,
-	Button,
-	Container,
-	Column,
-	Head,
-	Heading,
-	Hr,
-	Html,
-	Img,
-	Link,
-	Preview,
-	Row,
-	Section,
-	Tailwind,
-	Text,
-} from "@react-email/components";
-import type { DefaultEmailTemplateProps } from "@/lib/utils/server/types";
-import c from "config";
-
-interface RegistrationSuccessProps extends DefaultEmailTemplateProps {}
-
-const baseUrl = process.env.VERCEL_URL
-	? `https://${process.env.VERCEL_URL}`
-	: "";
-
-export default function RegistrationSuccessEmail({
-	firstName,
-}: RegistrationSuccessProps) {
-	return (
-		<div>
-			<Tailwind>
-				<Body className="mx-auto my-auto bg-white font-sans">
-					<Container className="mx-auto my-[40px] w-[465px] rounded border border-solid border-[#eaeaea] p-[20px]">
-						<Section className="mt-[32px]">
-							<Img
-								src={`${baseUrl}${c.icon.md}`}
-								width="40"
-								height="37"
-								alt={`${c.hackathonName} logo`}
-								className="mx-auto my-0"
-							/>
-						</Section>
-						<Heading className="mx-0 my-[30px] p-0 text-center text-[24px] font-normal text-black">
-							You are now registered for {c.hackathonName} $
-							{c.itteration}!
-						</Heading>
-						<Text className="text-[14px] leading-[24px] text-black">
-							Hello {firstName},
-						</Text>
-						{/* <Text className="text-black text-[14px] leading-[24px]">
-              <strong>bukinoshita</strong> (
-              <Link
-                href={`mailto:${invitedByEmail}`}
-                className="text-blue-600 no-underline"
-              >
-                {invitedByEmail}
-              </Link>
-              ) has invited you to the <strong>{teamName}</strong> team on{' '}
-              <strong>Vercel</strong>.
-            </Text>
-            <Section>
-              <Row>
-                <Column align="right">
-                  <Img className="rounded-full" src={userImage} width="64" height="64" />
-                </Column>
-                <Column align="center">
-                  <Img
-                    src={`${baseUrl}/static/vercel-arrow.png`}
-                    width="12"
-                    height="9"
-                    alt="invited you to"
-                  />
-                </Column>
-                <Column align="left">
-                  <Img className="rounded-full" src={teamImage} width="64" height="64" />
-                </Column>
-              </Row>
-            </Section>
-            <Section className="text-center mt-[32px] mb-[32px]">
-              <Button
-                pX={20}
-                pY={12}
-                className="bg-[#000000] rounded text-white text-[12px] font-semibold no-underline text-center"
-                href={inviteLink}
-              >
-                Join the team
-              </Button>
-            </Section>
-            <Text className="text-black text-[14px] leading-[24px]">
-              or copy and paste this URL into your browser:{' '}
-              <Link
-                href={inviteLink}
-                className="text-blue-600 no-underline"
-              >
-                {inviteLink}
-              </Link>
-            </Text>
-            <Hr className="border border-solid border-[#eaeaea] my-[26px] mx-0 w-full" />
-            <Text className="text-[#666666] text-[12px] leading-[24px]">
-              This invitation was intended for{' '}
-              <span className="text-black">{username} </span>.This invite was sent from{' '}
-              <span className="text-black">{inviteFromIp}</span> located in{' '}
-              <span className="text-black">{inviteFromLocation}</span>. If you were not
-              expecting this invitation, you can ignore this email. If you are
-              concerned about your account's safety, please reply to this email to
-              get in touch with us.
-            </Text> */}
-					</Container>
-				</Body>
-			</Tailwind>
-		</div>
-	);
-}
