+++ apps/web/src/lib/utils/server/types.ts
@@ -4,7 +4,3 @@ import type { NextResponse } from "next/server";
 export type serverZodResponse<T extends ZodType<any, any, any>> = Promise<
 	undefined | NextResponse<z.infer<T>> | NextResponse<"Unauthorized">
 >;
-
-export interface DefaultEmailTemplateProps {
-	firstName: string;
-}
