+++ apps/web/src/middleware.ts
@@ -35,5 +35,5 @@ export default clerkMiddleware(async (auth, req) => {
 });
 
 export const config = {
-	matcher: ["/((?!.*\\..*|_next).*)", "/", "/(api|trpc)(.*)"],
+	matcher: ["/((?!.*\\..*|_next).*)", "/", "/(api)(.*)"],
 };
