--- apps/web/src/server/api/root.ts
@@ -1,23 +1,0 @@
-import { ticketsRouter } from "@/server/api/routers/tickets";
-import { createCallerFactory, createTRPCRouter } from "@/server/api/trpc";
-
-/**
- * This is the primary router for your server.
- *
- * All routers added in /api/routers should be manually added here.
- */
-export const appRouter = createTRPCRouter({
-	tickets: ticketsRouter,
-});
-
-// export type definition of API
-export type AppRouter = typeof appRouter;
-
-/**
- * Create a server-side caller for the tRPC API.
- * @example
- * const trpc = createCaller(createContext);
- * const res = await trpc.post.all();
- *       ^? Post[]
- */
-export const createCaller = createCallerFactory(appRouter);
